Merge latest thunderbolt ICM from 01.org
Kernel 4.10 removed a couple of interfaces that the thunderbolt ICM module was making use of namely the `GENL_ID_GENERATE` macro and the `genl_register_family_with_ops` function. These are known userspace-breaks that happened on the 4.10 kernel and intel has already made fixes to accomidate them. This commit updates the driver to version 16.1.56.1 which includes those fixes.